+++ apps/web/frontend/userOverview.tsx
@@ -1,8 +1,24 @@
 import Image from 'next/image'
 import { useRouter } from 'next/router'
+import { useForm } from 'react-hook-form'
 
-import { Role, useMeQuery, UserFragment } from './generated/graphql'
+import { Button, InputField } from '@progwise/timebook-ui'
 
+import { Toggle } from './components/toggle/toggle'
+import {
+  Role,
+  useMeQuery,
+  useProjectMembershipCreateMutation,
+  useProjectMembershipDeleteMutation,
+  useTeamProjectsQuery,
+  useUserCapacityUpdateMutation,
+  useUserRoleUpdateMutation,
+  UserFragment,
+} from './generated/graphql'
+
+interface UserDetailsForm {
+  availableMinutesPerWeek: string
+}
 interface UserOverviewProps {
   user: UserFragment
 }
@@ -10,8 +26,44 @@ interface UserOverviewProps {
 export const UserOverview = ({ user }: UserOverviewProps): JSX.Element => {
   const router = useRouter()
   const teamSlug = router.query.teamSlug?.toString() ?? ''
+  const [{ data: allProjects }] = useTeamProjectsQuery({ variables: { slug: teamSlug } })
   const [{ data: meData }] = useMeQuery({ variables: { teamSlug } })
+  const [, createProjectMembership] = useProjectMembershipCreateMutation()
+  const [, deleteProjectMembership] = useProjectMembershipDeleteMutation()
   const isAdmin = meData?.user.role === Role.Admin
+
+  const [{ error, fetching }, userRoleUpdate] = useUserRoleUpdateMutation()
+  const [, updateUserCapacity] = useUserCapacityUpdateMutation()
+
+  const submitHandler = async (data: UserDetailsForm) => {
+    const response = await updateUserCapacity({
+      // eslint-disable-next-line unicorn/no-null
+      availableMinutesPerWeek: data.availableMinutesPerWeek.length > 0 ? +data.availableMinutesPerWeek : null,
+      userId: user.id,
+      teamSlug,
+    })
+
+    if (response.error) setError('availableMinutesPerWeek', { message: response.error.message })
+  }
+
+  const handleUpgradeClick = () => {
+    userRoleUpdate({ role: Role.Admin, userId: user.id, teamSlug })
+  }
+
+  const handleDowngradeClick = () => {
+    userRoleUpdate({ role: Role.Member, userId: user.id, teamSlug })
+  }
+
+  const {
+    register,
+    formState: { errors: fieldsErrors },
+    setError,
+    handleSubmit,
+  } = useForm<UserDetailsForm>({
+    mode: 'onChange',
+    defaultValues: { availableMinutesPerWeek: user.availableMinutesPerWeek?.toString() },
+  })
+
   return (
     <>
       <article>
@@ -28,9 +80,9 @@ export const UserOverview = ({ user }: UserOverviewProps): JSX.Element => {
               <h1 className="text-xl font-semibold text-gray-400">Team member</h1>
 
               <div className="flex">
-                {isAdmin && meData?.user.id !== userId && (
+                {isAdmin && meData?.user.id !== user.id && (
                   <>
-                    {data?.user.role === Role.Member ? (
+                    {user.role === Role.Member ? (
                       <Button className="mr-4" variant="secondary" onClick={handleUpgradeClick}>
                         Promote to admin
                       </Button>
@@ -56,15 +108,12 @@ export const UserOverview = ({ user }: UserOverviewProps): JSX.Element => {
                   <li key={project.id} className="p-3">
                     <span className=" inline-block w-32"> {project.title} </span>
                     <Toggle
-                      checked={data?.user.projects.some((userProject) => userProject.id === project.id) ?? false}
+                      checked={user.projects.some((userProject) => userProject.id === project.id) ?? false}
                       onChange={(newValue) => {
-                        if (!data?.user) {
-                          return
-                        }
                         if (newValue === false) {
-                          deleteProjectMembership({ projectID: project.id, userID: data?.user.id })
+                          deleteProjectMembership({ projectID: project.id, userID: user.id })
                         } else {
-                          createProjectMembership({ projectID: project.id, userID: data?.user.id })
+                          createProjectMembership({ projectID: project.id, userID: user.id })
                         }
                       }}
                     />
@@ -102,7 +151,7 @@ export const UserOverview = ({ user }: UserOverviewProps): JSX.Element => {
           <>
             <h1 className="text-xl font-semibold text-gray-400"> Assigned Projects:</h1>
             <ul>
-              {data?.user.projects.map((project) => (
+              {user.projects.map((project) => (
                 <li key={project.id}> {project.title}</li>
               ))}
             </ul>
